requirements.txt
yt-dlp
python-aalib
Pillow
